Rename Nix flake package
The `-node` part was added while I tried to make the whole thing work,
and was considering having a package for source and modules and another
for the output.  Since I figured out how to do it all in a single
package, the suffix is no longer necessary nor correct.